sevencontinent/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg[binary]
python-dotenv==1.0.0
Pillow
django-cors-headers==4.0.0
drf-yasg>=1.21.8
django-filter
gunicorn==22.0.0
